+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- feat(nextjs): Support all `next.config` file types (#630)
+- fix(nextjs): Update instrumentation and example creation logic for app or pages usage (#629)
+
 ## 3.25.2
 
 - ref: Improve intro and wizard selection (#625)